// src/types/language.ts
export type SupportedLanguageCode = 
  | 'en' | 'ta' | 'hi' | 'te' | 'kn' | 'ml' | 'mr' | 'bn' 
  | 'gu' | 'pa' | 'or' | 'as' | 'ur' | 'sa' | 'ks' | 'ne' 
  | 'kok' | 'mai' | 'mni' | 'brx' | 'doi' | 'sat' | 'sd';